Merge testing into master

Former-commit-id: b7db927c566769a15fe5f7d5fbaccc63cfba8911

+++ src/server/game/playerbot/strategy/actions/BuffAction.cpp
@@ -93,7 +93,7 @@ bool BuffAction::Execute(Event event)
     FindBuffVisitor visitor(bot);
     IterateItems(&visitor);
 
-    uint32 oldSubClass = -1;
+    uint32 oldSubClass = uint32(-1);
     for (map<uint32, list<Item*> >::iterator i = visitor.items.begin(); i != visitor.items.end(); ++i)
     {
         list<Item*> items = i->second;
